[CHK-702] Update currency format, optimize imports (#20)
* [CHK-702] fix: update currency format, optimize imports

* [CHK-702] fix: optimize imports

* [CHK-702] fix: update tests

+++ src/main/java/it/pagopa/ecommerce/payment/methods/domain/valueobjects/PspFee.java
@@ -4,20 +4,21 @@ import lombok.EqualsAndHashCode;
 import org.springframework.lang.NonNull;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.Objects;
 
 @ValueObjects
 @EqualsAndHashCode
 public class PspFee implements Serializable {
 
-    private final Double fee;
+    private final BigInteger fee;
 
-    public PspFee(@NonNull Double fee) {
+    public PspFee(@NonNull BigInteger fee) {
 
         this.fee = Objects.requireNonNull(fee);
     }
 
-    public @NonNull Double value() {
+    public @NonNull BigInteger value() {
         return fee;
     }
 }
